Added the windows alert popup

+++ src/app/main-nav/main-nav.component.ts
@@ -122,11 +122,13 @@ export class MainNavComponent {
   }
 
   deleteApplicationType(id: any) {
-    this.applicationService.deleteApplicationType(id).subscribe (
-      response => {
-        console.log(response);
-        this.retrieveApplicationTypes();
-      }
-    );
+    if(window.confirm('Are sure you want to delete this item ?')){
+      this.applicationService.deleteApplicationType(id).subscribe (
+        response => {
+          console.log(response);
+          this.retrieveApplicationTypes();
+        }
+      );
+    }
   }
 }
